+++ src/Core/Geometry/Adjacency/Adjacency.hpp
@@ -6,7 +6,7 @@
 #include <Core/Math/LinearAlgebra.hpp>
 #include <Core/Geometry/MeshTypes.hpp>
 
-#include <Core/Utils/Graph/AdjacencyList.hpp>
+#include <Core/Utils/AdjacencyList.hpp>
 
 namespace Ra {
 namespace Core {
