+++ src/Core/Utils/TaskQueue.cpp
@@ -1,5 +1,5 @@
-#include <Core/Tasks/Task.hpp>
-#include <Core/Tasks/TaskQueue.hpp>
+#include <Core/Utils/Task.hpp>
+#include <Core/Utils/TaskQueue.hpp>
 
 #include <algorithm>
 #include <iostream>
