+++ src/Engine/Managers/LightManager/LightManager.cpp
@@ -5,8 +5,8 @@
 
 #include <Core/Asset/FileData.hpp>
 #include <Core/Asset/GeometryData.hpp>
-#include <Core/Tasks/Task.hpp>
-#include <Core/Tasks/TaskQueue.hpp>
+#include <Core/Utils/Task.hpp>
+#include <Core/Utils/TaskQueue.hpp>
 
 #include <Engine/Entity/Entity.hpp>
 #include <Engine/FrameInfo.hpp>
